rename everything from js to ts(x)
